refactor: base installation scripts
install from GH if no packages found

+++ .assets/provision/install_bat.sh
@@ -45,10 +45,11 @@ debian | ubuntu)
 opensuse)
   zypper in -y bat >&2
   ;;
-*)
+esac
+
+if ! type bat &>/dev/null; then
   while [[ ! -d "bat-v${REL}-x86_64-unknown-linux-gnu" ]]; do
     curl -Lsk "https://github.com/sharkdp/bat/releases/download/v${REL}/bat-v${REL}-x86_64-unknown-linux-gnu.tar.gz" | tar -zx
   done
   install -o root -g root -m 0755 "bat-v${REL}-x86_64-unknown-linux-gnu/bat" /usr/bin/bat && rm -fr "bat-v${REL}-x86_64-unknown-linux-gnu"
-  ;;
-esac
+fi
